PROBE Move code
